Default to vertical Sony profile if the video is vertical

+++ src/core/gpu/opencl_undistort.cl
@@ -318,7 +318,7 @@ float2 rotate_and_distort(float2 pos, uint idx, __global KernelParams *params, _
             float2 origin    = (float2)(lens_data[4], lens_data[5]);
             float2 crop_size = (float2)(lens_data[6], lens_data[7]);
 
-            if (params->flags & 128) uv.y = params->height - uv.y; // framebuffer inverted
+            if (params->flags & 128) uv.y = (float)params->height - uv.y; // framebuffer inverted
 
             uv.x = map_coord(uv.x, 0.0f, (float)params->width,  origin.x, origin.x + crop_size.x);
             uv.y = map_coord(uv.y, 0.0f, (float)params->height, origin.y, origin.y + crop_size.y);
@@ -328,7 +328,7 @@ float2 rotate_and_distort(float2 pos, uint idx, __global KernelParams *params, _
             uv.x = map_coord(uv.x, origin.x, origin.x + crop_size.x, 0.0f, (float)params->width);
             uv.y = map_coord(uv.y, origin.y, origin.y + crop_size.y, 0.0f, (float)params->height);
 
-            if (params->flags & 128) uv.y = params->height - uv.y; // framebuffer inverted
+            if (params->flags & 128) uv.y = (float)params->height - uv.y; // framebuffer inverted
         }
 
         if (params->flags & 2) { // Has digital lens
